fixup! Move from Snackbar to Alert for coloring
We discovered bugs with our submission
notification code. There is incosistency in
whether or not the browser will render the
CSS specified by 'BBFL/src/components/NFL/NFL.css'.

I couldn't find a way to reliably change the
snackbar component color. In turn, I switched
to using the Alert as the mutable component.
This seems to work well.

There is no issue to track this work.