fixup! Cleaning up ESPN's diaper
ESPN has not updated the season type
in their API. Furthermore, they seem to have
removed the element noting the current week number.
In turn, the espn-api-parser is failing for both of the above.

This hash hard-codes the week number to one if the week_number
element does not exist. It also enforces the use of ESPN's own
understanding of the current week of the season.

Testing: Manual