feat! Adds google login.
It's currently rather complicated to resolve
password issues as we don't currently support
"Forgot my password" situations.

This hash resolves this by pawning the auth
off to supabase in google, which gives us one
less thing to have to program and maintain. It
also locks down access to individual pages and
switches to BrowserRouter over that nasty
HashRouter.