fixup! Late pick detection bug
Any pick passed Thursday is being marked as late because that
match has technically ended.

This hash introduces ugly logic to re-fetch the user's picks
as stored in the db (source of truth), and checks to see which
picks are different. In the future, we can optimize our submit
picks algo by excluding the picks that aren't being updated.
But for now, I'm lazy.

We should also be storing the user's db picks in the heap so we don't
have to re-fetch them at all.